pyproject.toml
[project]
name = "finance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.3.5",
    "ddgs>=9.9.3",
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.124.0",
    "langchain>=0.1.0",
    "langchain-community>=0.2.10,<0.3.0",
    "langchain-huggingface>=0.0.3",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "uvicorn>=0.38.0",
    "yfinance>=0.2.66",
]
